requirements.txt
streamlit
langchain
langchain-google-genai
langchain-core
python-dotenv